feat: 컨텐츠 상세 페이지 탭 추가
