+++ app/(home)/content/page.tsx
@@ -8,37 +8,22 @@ import IconX from '@/assets/images/icon_x.svg';
 import RegisterBtn from '@/components/RegisterBtn';
 import { catIllust } from '@/constants/catIllust';
 import { pinList } from '@/constants/contentMockData';
-import { Label } from '@/components/Input';
-import {
-  groupButtons,
-  neuterButtons,
-  personalityButtons,
-} from '@/constants/catInfoButtons';
-import MapComponent from '@/components/Map/Map';
-import CustomPin from '@/components/Map/CustomPin';
+import { DetailInfo } from '@/components/Content/DetailInfo';
 
 export interface ContentType {
   id: number;
-  lat: number;
-  lng: number;
-  nickname: string;
-  name: string;
-  neuter: string;
-  group: string;
-  personality: string[] | [];
-  desc: string;
-  createdAt: string;
-  addrName: string;
-  comment: number;
-  follow: number;
-  catEmoji: number;
 }
+
+const tabName = [
+  { name: '상세 정보', value: 'detailInfo' },
+  { name: '냥이 소식', value: 'catNews' },
+];
 
 const RegisterPostPage = () => {
   const router = useRouter();
   const content = pinList[0];
   const cat = catIllust.filter((cat) => cat.id === content.catEmoji)[0];
-  const position = { lat: content.lat, lng: content.lng };
+  const [tab, setTab] = useState('detailInfo');
 
   return (
     <Fragment>
@@ -71,74 +56,28 @@ const RegisterPostPage = () => {
           </div>
         </div>
 
-        <div className='p-6'>
-          <div className='caption2 text-primary-400 mb-3'>
-            {content.updatedAt} 업데이트
-          </div>
-          <div className={`${titleClassName}`}>동네 집사의 한 줄 소개</div>
-          <div className='body2 text-gray-400'>{content.desc}</div>
-        </div>
-
-        <div className={`${barClassName}`} />
-
-        <div className='p-6'>
-          <div className={`${titleClassName}`}>
-            건강한 돌봄을 위한 필수 체크!
-          </div>
-          <div className='flex mb-4'>
-            <div className='w-full'>
-              <div className={`${subTitleClassName}`}>중성화 수술</div>
-              <div>
-                <span className={`${contentClassName}`}>
-                  {
-                    neuterButtons.find(
-                      (button) => button.value === content.neuter,
-                    )?.name
-                  }
-                </span>
-              </div>
-            </div>
-            <div className='w-full'>
-              <div className={`${subTitleClassName}`}>같이 다니는 무리</div>
-              <div>
-                <span className={`${contentClassName}`}>
-                  {
-                    groupButtons.find(
-                      (button) => button.value === content.group,
-                    )?.name
-                  }
-                </span>
-              </div>
-            </div>
-          </div>
-
-          <div className={`${subTitleClassName}`}>성격 및 특징</div>
-          <div className='flex flex-wrap'>
-            {content.personality.map((value) => (
+        <div>
+          <div className='flex px-6 mb-[10px]'>
+            {tabName.map((v) => (
               <div
-                key={value}
-                className={`${contentClassName} mr-[6px] text-nowrap mb-[6px]`}
+                className={`w-full text-center cursor-pointer ${v.value === tab ? 'text-gray-500' : 'text-gray-200'}`}
+                onClick={() => setTab(v.value)}
               >
-                {
-                  personalityButtons.find((button) => button.value === value)
-                    ?.name
-                }
+                {v.name}
               </div>
             ))}
           </div>
-        </div>
-
-        <div className={`${barClassName}`} />
-
-        <div className='p-6'>
-          <div className={`${titleClassName}`}>주요 출몰 위치</div>
-          <div className='caption text-gray-400 mb-3'>{content.addrName}</div>
-          <div className='w-full h-[136px]'>
-            <MapComponent position={position} level={4}>
-              <CustomPin position={position} />
-            </MapComponent>
+          <div className='flex px-6'>
+            {tabName.map((v) => (
+              <div
+                className={`w-full h-[1.5px] z-10 ${v.value === tab ? 'bg-gray-500' : 'bg-gray-50'}`}
+              />
+            ))}
           </div>
+          <div className='absolute w-full h-[1.5px] bg-gray-50 translate-y-[-1.5px]' />
         </div>
+
+        {tab === 'detailInfo' ? <DetailInfo content={content} /> : null}
       </div>
 
       <div className='absolute bottom-0 left-0 w-full z-20 px-6 pt-[18px] pb-[30px] shadow-[0px_-8px_8px_0px_rgba(0,0,0,0.15)] bg-white'>
@@ -151,9 +90,3 @@ const RegisterPostPage = () => {
 };
 
 export default RegisterPostPage;
-
-const titleClassName = 'flex flex-col pb-2 subHeading';
-const subTitleClassName = 'subHeading2 text-gray-400 mb-2';
-const contentClassName =
-  'caption text-primary-500 bg-primary-100 px-[10px] py-[6px] rounded';
-const barClassName = 'w-full h-[7px] bg-gray-50';
